Move types to Types collection file
